requirements.txt
streamlit
python-dotenv
PyPDF2
requests
sentence-transformers
faiss-cpu
numpy
google-genai
certifi
charset-normalizer
idna
urllib3